# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0", 
    "title": "Manager app backend",
    "description": "Manager Application API",
    "license": "ISC"
  },
...
  "securityDefinitions":{
    "Bearer":{
      "type": "apiKey",
      "name": "Authorization",
      "description": "Enter your token in format *Bearer token*",
      "in": "header"
    }
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths":{
...
    "/boards/{boardId}/columns": {
      "get": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Columns"],
        "summary": "Get Columns in Board",
        "description": "Get Columns in board",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "All columns in board",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Column"
              } 
            }
          }
        }
      },
      "post": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Columns"],
        "summary": "Create Column",
        "description": "Create Column in board",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "new Column params",
            "in": "body",
            "description": "New params",
            "schema": {
              "required": [
                "title",
                "order"
              ],
              "properties": {
                "title": {
                  "type": "string"
                },
                "order": {
                  "type": "number"
                }      
              }
            }
            
          }
        ],
        "responses": {
          "200": {
            "description": "Created column",
            "schema": {
              "$ref": "#/definitions/Column"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"              
            }
          }
        }
      }
    },    
    "/boards/{boardId}/columns/{columnId}": {
      "get": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Columns"],
        "summary": "Get Column by Id",
        "description": "Find Column",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Founded column",
            "schema": {
              "$ref": "#/definitions/Column"
            }
          },
          "404": {
            "description": "Column was not founded!",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Columns"],
        "summary": "Update Column by Id",
        "description": "Update Column",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          },          
          {
            "name": "new Column params",
            "in": "body",
            "description": "New params",
            "schema": {
              "required": [
                "title",
                "order"
              ],
              "properties": {
                "title": {
                  "type": "string"
                },
                "order": {
                  "type": "number"
                }       
              }
            }
            
          }
        ],
        "responses": {
          "200": {
            "description": "Updated column",
            "schema": {
              "$ref": "#/definitions/Column"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"              
            }
          }
        }
      },
      "delete": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Columns"],
        "summary": "Delete Column by Id",
        "description": "Delete Column",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted column",
            "schema": {
              "$ref": "#/definitions/Column"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Column": {
      "required": [
        "_id",
        "title",
        "order",
        "boardId"
      ],
      "properties": {
        "_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "order": {
          "type": "number"
        },
        "boardId": {
          "type": "string"
        }      
      },
      "example": {
        "_id": "Column id",
        "title": "Column title",
        "order": 1,
        "boardId": "Id of boards"
      }
    },
...
    "ErrorResponse": {
      "required": [
        "statusCode",
        "message"
      ],
      "properties": {
        "statusCode": {
          "type": "number"
        },
        "message": {
          "type": "string"
        }     
      },
      "example": {
        "statusCode": "4xx",
        "message": "reason"
      }
    }
  }
}